Basic Level Elements
Setup the Debug Map to have some of the artwork made by the wonderful Tiabwwtws just to test. Scroll Camera works as intended, it just may need a few tweaks.

Made some tweaks to the movement speed on the move state as it felt like I was going to a snail's pace.

We'll just need the puzzle elements.

Light Wizard actually shooting light soon :tm: