Big Commit - Read Desc
Just adds abunch of the stuff that I've been working on.

Learned that having the initial resolution being lower is kinda important. We can scale it later to our liking, but it has to initially be lower.

Lever can be switched. It will also be able to be switched by the player's light projectile.
Stairs can be climbed by the player.
Mirror is in progress. Player will be able to turn it and reflect light off of it as needed.
Removed redundant objects/scenes.
Changed a few states.
Added Sound to Cloud's menu buttons along with comments.
Gave the WorldMap the 3rd layer for the projectile to detect a collision there and disappear accordingly.